:sparkles: enable taking http proxy settings (#230)
Fixes #101 

1. Adds a new `proxyConfig` in provider settings. This is so that
providers know whether a proxy should be used. This is technically only
useful for external providers. Right now, we only have go provider which
is basically taking the provider settings and setting proxy config as
environment variables because `go` respects standard system settings.
Other providers, however, may use these settings differently.
2. When a specific config is not specified, system-wide proxy settings
set via env vars are respected. Everything running in the main process
respects system-wide settings, thanks to golang.
3. Maven running within JDTLS, however, does not respect system
settings. This should be a bug IMO because `java.net.useSystemProxy` is
a env var that most Java programs respect, which m2e core doesn't seem
to respect. For this case, we need to specify settings via
`settings.xml` file. This file is a common practice within users to set
proxy settings for maven, we also have support for taking a settings
file via Hub. As a result, we are adding a `providerSpecificConfig`
field for Java that allows taking this file in addition to the normal
proxy settings.


*NOTE*: There is one case which I haven't tested - when an external
provider runs outside the network of the main process, and there is a
proxy to establish connection between the provider and the rule engine
itself. I believe that gRPC client should respect system proxies in that
case too and we don't need to do anything special there. I am not sure
right now. I am creating an issue just to make sure we don't lose track
of that case.

Signed-off-by: Pranav Gaikwad <pgaikwad@redhat.com>