:sparkles: Refactors code that deals with condition response and violations (#298)
This pull request addresses issue #235 .
Removes unused code that was redundant as it's functionality was handled
by other elements. Actually, the `links` field in `IncidentContext`
draws it's value from `rule`, and as `rule` is also passed to
`createViolations` function, we can derive the `links` in the function
itself. So, the `links` field within `IncidentContext` is currently
unused throughout the source code; its primary role was to transfer this
value to `violations`, which is handled by `rule`. With your approval,
we can safely remove this field to enhance code clarity.
@pranavgaikwad please check

Signed-off-by: AdiAkhileshSingh15 <adiakhilesh.singh.cse21@itbhu.ac.in>