:ghost: only decompile when directory doesnt exist (#336)
Since we moved the decompilation into filter.go, we are now decompiling
prior to filtering the incidents out. In that case, for files with
.groovy, .kotlin etc, we are decompiling unnecessarily multiple times
without knowing that we are anyway going to fail getting incidents from
them later on. This was causing some shim tests to take awful lot of
time.

Signed-off-by: Pranav Gaikwad <pgaikwad@redhat.com>