:bug: Use qaware's dependency plugin to avoid all issues related to maven dependency plugin (#426)
API Tests PR: 55

With the new plugin, dependency extraction is simplified to parsing a
single line. We also don't need to download dependencies and sources
separately; everything is done in a single command. If a dependency is
not found in any of the configured repositories, then we won't be able
to decompile it (it is absent), therefore we are not interested in those
and ignore them.

- Fixes https://github.com/konveyor/analyzer-lsp/issues/423
- We also wouldn't need
https://github.com/konveyor/analyzer-lsp/pull/419 anymore

Signed-off-by: Juan Manuel Leflet Estrada <jleflete@redhat.com>