:sparkles: add dependency location resolver and get code locations for XML, JSON files  (#412)
Fixes #410 
Fixes #254 

There were two choices to do this: 

1. Add a new optional field lineNumber to Dep type, and let the
providers populate it at the time of dependency discovery. This is
straight-forward and probably preferred, but requires a change in the
output API.
2. Add a new interface like Snipper, that providers can implement for
deps to get locations for a certain dependency.

I chose No. 2 because output API change not needed. But I am not fan of
it as its complicated & perf penalty. Asking for your inputs on the
approach. If we do go with 2, I think we should we should still remove
it in later versions in favor of a new optional field on Dep. Even if we
did add it now, it would be non-intrusive.

---------

Signed-off-by: Pranav Gaikwad <pgaikwad@redhat.com>