:sparkles: Add dockerfiles for providers and run the demo using the external providers in a Pod (#579)
Fixes #470 

I have still kept the original base Dockerfile as-is for backwards
compatibility. This is the first step in which we are chaging our demo
tests to use the external providers. I will create a follow up issue to
track the work around actually making the switch in the addon.

---------

Signed-off-by: Pranav Gaikwad <pgaikwad@redhat.com>
Co-authored-by: David Zager <dzager@redhat.com>