:sparkles: Make providers return a builtin config so engine can look at updated locations  (#671)
Some providers can mutate the original location given in init config.
For example, for binaries, java provider decompiles and uses the new
decompiled path. This PR adds a way for providers to communicate updated
locations back to the engine so builtin provider can be configured
correctly without the user having to do it.

---------

Signed-off-by: Pranav Gaikwad <pgaikwad@redhat.com>
Signed-off-by: Shawn Hurley <shawn@hurley.page>
Co-authored-by: Pranav Gaikwad <pgaikwad@redhat.com>
Signed-off-by: Cherry Picker <noreply@github.com>