:sparkles: try to get import references when found (#688)
@djzager

Basically, when we do our fall back we can find the imports mostly
correct, but imports are for namespaces not for actual objects so no
definition can be used. Therefore we are getting zero results back for
the definiton.

We can get all the references for the import though, and we can just let
the analyzer dedup.

If you could help get me a windows container image to test with that
would helpful (or show me how)

Signed-off-by: Shawn Hurley <shawn@hurley.page>
Signed-off-by: midays@redhat.com <midays@redhat.com>