Log full Maven output on errors
Java provider uses mvn command to fetch java binary or its dependencies.
There was a check for success/error of those commands, but to identify a
root cause and fix the mvn problem, its full output is required.

The maven output can be be exhaustive, but it is logged only in case of
its failure.

Signed-off-by: Marek Aufart <maufart@redhat.com>