:bug: Add cancel context to NewGRPCClient (#720)
### PR Summary
Bug fix to address the issue of java-external-provider.exe not ending
after analysis completion.
Introduces a cancel context to the NewGRPCClient to ensure proper
termination of the gRPC client.

- provider/grpc/provider.go: Added context.WithCancel to manage the
lifecycle of the gRPC client and ensure it can be properly stopped.
- provider/grpc/provider.go: Updated grpcProvider struct to include
cancelCmd and invoked cancelCmd() in the Stop method to terminate the
context.

Signed-off-by: kthatipally <99229993+kthatipally@users.noreply.github.com>
Signed-off-by: Cherry Picker <noreply@github.com>