Use XML line numbers directly from xmlquery
There is a problem with line number identification for XML matched nodes
(empty node, just with attributes and could be spanning multiple lines
in the source XML file). This lead to find all relevant matches using
xpath query, however all had the same incident line number, so we
deleted as duplicate during the analysis.

The fix of the root cause of the issue is to set correct location of
the matched element. It was complicated/close to impossible IMO to do it
in reliable way with just update of current methods in analyzer.

In order to have XML Nodes correctly annotated with their line numbers
in the source file, the `xmlquery` package was forked and updated. We
should confirm if this is an acceptable way and potentialy try get this
change into official upstream `antchf/xmlquery` repository, changes:
https://github.com/antchfx/xmlquery/compare/master...aufi:xmlquery:line-numbers?expand=1

Fixes https://github.com/konveyor/analyzer-lsp/issues/795

Signed-off-by: Marek Aufart <maufart@redhat.com>