:wrench: Add replace directive to java provider go.mod to fix CI
The replace directive is required for CI to pass because this PR includes
both the NormalizePathForComparison function (added in commit 5bbe00a) and
its usage in the java provider. Without the replace directive, go.mod
references the older v0.9.0-alpha.1.0.20251121202139-b3f742caf596 version
which doesn't have this function, causing build failures.

This can be removed in a follow-up PR after merge.

Signed-off-by: tsanders <tsanders@redhat.com>