:ghost: Adding ability to rebuild the java-provider when the jdtls server asks (#1022)
Currently, you have to manually know to rebuild the analyzer images to
pick up the latest changes in the jdtls-server-base. This has caused
alot of issues with debugging CI. This will enable the
java-analyzer-bundle to ask for the java-provider to be rebuilt.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

* **Chores**
* Added CI/CD workflow to build and publish multi-architecture Java
external-provider container images (amd64, arm64), enabling automated
image builds and tagging for releases and branch refs.

<sub>✏️ Tip: You can customize this high-level summary in your review
settings.</sub>
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Signed-off-by: Shawn Hurley <shawn@hurley.page>