:recycle: Add ChainTemplate.ToMap() helper method
Add a ToMap() helper method to the ChainTemplate struct to make
the map creation logic more maintainable. This refactors the manual
map creation in templateCondition() to use the new helper, making it
more straightforward to see that a change to this logic is needed if
a field is ever added to ChainTemplate.

As suggested by Shawn's feedback.

Signed-off-by: tsanders <tsanders@redhat.com>