:bug: Always create throttledReporter for GRPC streaming support
For GRPC providers, PrepareProgressReporter cannot be passed through
the Init RPC as it's not serializable. However, the throttledReporter
is still needed to enable progress event streaming.

This change ensures throttledReporter is always created, using a no-op
reporter when PrepareProgressReporter is nil. This allows GRPC providers
to stream progress events even without a direct reporter.

Signed-off-by: tsanders <tsanders@redhat.com>